# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API for El Diego e-commerce application",
        "contact": {
            "email": "diegomalave@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/v1.0.0/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List active products (paginated)",
                "description": "Display a paginated list of active products.",
                "operationId": "c8ba846e2886ff66ff7c5002b79230fb",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of products per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of active products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "ssd"
                                                    },
                                                    "sku": {
                                                        "type": "string",
                                                        "example": "SKU-22276"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "333.00"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "11111111111111111",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-23T13:06:36+00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-23T13:42:33+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://e-commerce.test/api/v1.0.0/products?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://e-commerce.test/api/v1.0.0/products?page=4"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://e-commerce.test/api/v1.0.0/products?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://e-commerce.test/api/v1.0.0/products"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 52
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Store a newly created product in the database.",
                "operationId": "f1cbc7a78d47d0b726e0c3ead8b5d3aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sku",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "product 44"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-345345"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "333.23"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "adl sdh sdfh sdfuk sd k dfkL SDF  SDFKH SDKFD",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The product has been successfully created."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 79
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "product 44"
                                                },
                                                "sku": {
                                                    "type": "string",
                                                    "example": "SKU-345345"
                                                },
                                                "price": {
                                                    "type": "string",
                                                    "example": "333.23"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "adl sdh sdfh sdfuk sd k dfkL SDF  SDFKH SDKFD"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-23T18:14:44+00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-23T18:14:44+00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The sku has already been taken."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}